requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
scikit-learn
numpy
matplotlib
seaborn
jinja2